feat: show documentation popup in completion
